+++ Extensions.Graph/VersionHistory.md
@@ -3,6 +3,9 @@
 ### **4.9.700 (2023-03-29)**<br>
 	- Separate `Extensions.Graph` class from Extensions.cs.
 
-### **6.0.800 (2024-02-27)**<br>
+### **6.0.800 (2024-02-26)**<br>
 	- Rearchitected and Optimized.<br>
 	- Rebranded the `Extensions` set of NuGet packages as `GCCHigh.Extensions`.
+
+### **6.1.800 (2024-02-27)**<br>
+	- Added the `getFile` option to `.GetDriveItem()` to allow for retrieval of the underlying binary file.