+++ Extensions.Identity/VersionHistory.md
@@ -7,14 +7,17 @@
 	- Rearchitected and Optimized.<br>
 	- Rebranded the `Extensions` set of NuGet packages as `GCCHigh.Extensions`.
 
-### **6.1.800 (2024-02-27)**<br>
+### **6.1.800 (2024-02-26)**<br>
 	- Change .LoadConfig() from internal to public.
 
-### **6.2.800 (2024-02-27)**<br>
+### **6.2.800 (2024-02-26)**<br>
 	- Added `ActiveAuth` validation to the `.GetAuth()` method.
 	- Added `ActiveAuth` validation to the `.GetHttpClient()` method.
 	- Added `ActiveAuth` validation to the `.GetGraphServiceClient()` method.
 	- Added `ActiveAuth` validation to the `.GetGraphBetaServiceClient()` method.
 
-### **6.3.800 (2024-02-27)**<br>
+### **6.3.800 (2024-02-26)**<br>
 	- Added `ActiveAuth` validation to the `TenantConfig` properties Get methods.
+
+### **6.4.800 (2024-02-27)**<br>
+	- Added the `TargetTenantConfig` to AuthMan for initialization optimization.