+++ VersionHistory.md
@@ -387,6 +387,11 @@
 	- Updated Microsoft.Graph dependency to 5.42.0.<br>
 	- Updated Microsoft.Identity.Client dependency to 4.59.0.<br>
 
-### **6.0.800 (2024-02-27)**<br>
+### **6.0.800 (2024-02-26)**<br>
 	- Rearchitected and Optimized.<br>
 	- Rebranded the `Extensions` set of NuGet packages as `GCCHigh.Extensions`.
+
+### **6.1.800 (2024-02-27)**<br>
+	- Optimized tenant initialization logic in Core.
+	- Added the `getFile` option to `Graph.GetDriveItem()` to allow for retrieval of the underlying binary file.
+	- Added the `TargetTenantConfig` to Identity for initialization optimization.